Here is a clean, conventional commit message summarizing the changes in **commit_context.txt** :
---

**refactor: split monolithic Gutenberg Query Builder UI into modular package**

* Replace `scripts/production/gutenberg_query_builder_ui.py` (1567 lines) with a fully modular package:

  * `__init__.py` – public exports + `main()`
  * `__main__.py` – module execution entry point
  * `app.py` – main application class
  * `constants.py` – configuration + operator/field definitions
  * `tk_helpers.py` – isolated tkinter/pandas wrappers with type suppressions
  * `widgets.py` – tooltip, constraint widgets, group widgets
* Update VS Code launch configuration to use module execution (`python -m scripts.production.gutenberg_query_builder_ui`)
* Update README and usage instructions to match new execution model
* Add documentation: `docs/refactoring-gutenberg-query-builder-ui.md` detailing the full refactor
* Update and realign `test_query_builder_ui.py` to the new package structure

**Outcome**

* Clean package-based architecture
* Zero behavioral changes; all 466 tests still pass
* Pyright-clean (0 errors), Ruff-clean, Black-formatted
* Improved maintainability, separation of concerns, and type-safety

---
